refactor: reground components
