+++ src/navigation/components/AnimatedSplash/index.tsx
