+++ src/screens/Team/components/Contacts/components/ModifyPhoneNumber/hooks/useModifyPhoneNumberWalkthrough.tsx
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: ice License 1.0
 
+import {PrimaryButton} from '@components/Buttons/PrimaryButton';
 import {UPDATE_ACCOUNT_FIELD_BUTTON_OFFSET} from '@components/Forms/components/UpdateAccountField';
 import {PhoneNumberInput} from '@components/Inputs/PhoneNumberInput';
-import {PrimaryButton} from '@components/PrimaryButton';
 import {SCREEN_SIDE_OFFSET} from '@constants/styles';
 import {WalkthroughElementContainer} from '@screens/Walkthrough/components/WalkthroughElementContainer';
 import {useSetWalkthroughElementData} from '@store/modules/Walkthrough/hooks/useSetWalkthroughElementData';
